refactor: extract risk filter logic
