+++ quant_trade/signal/__init__.py
@@ -38,6 +38,7 @@ from .utils import (
 from .voting_model import VotingModel, load_cached_model
 from .factor_scorer import FactorScorerImpl
 from .fusion_rule import FusionRuleBased
+from .risk_filters import RiskFiltersImpl
 
 __all__ = [
     "SignalThresholdParams",
@@ -73,4 +74,5 @@ __all__ = [
     "load_cached_model",
     "FactorScorerImpl",
     "FusionRuleBased",
+    "RiskFiltersImpl",
 ]
